InDetEventAthenaPool: Avoid undefined behavior.
Code in this package was taking advantage of undefined behavior to change
the name used by the MsgStreams of converters (a protected data member)
to something more meaningful.  This results in warnings from the
undefined behavior sanitizer.

cnvname.AthenaPoolCnvSvc added an optional argument to the converter 
constructors to allow specifying this name.  So switch to using this.



Former-commit-id: 746f7f53292714b7ea77c54d94b233cd43c92946