Add functionality to check whether Pixel FE Chips are masked
Currently no explicit check is made on whether a FE chip is active, despite the information being available in conditions
These changes add a flag that allows FEs to be checked. This is done within the function that checks the fraction
of good channels in a modules, in the region related to the track uncertainty. The changes as is maintina frozen Tier0 since
the FE checking is not enable by default.


Former-commit-id: ce0258168b46b36cca9914d84fbfb0dc739ec22b