Replacing #defines for T/P Converters with typedefs
Replaced #defines for T/P Converters that were needed after the conveter
move to the new package TPTools and name changes, with typdefs. That was
always intended but did not work with GCCXML so a #define hack was used.
Also cleaned up the associated #includes.
Conveter names on the user side were not renamed this time.


Former-commit-id: 0ddcc8724bbd57169b2cc2ab8c135ba1c3a6d965