Fix for ATLASRECTS-3796 plus smaller technical changes
This change should fix ATLASRECTS-3796 by adding dummy truth links to xAOD::Muon objects and muon tracks that are not truth-matched, thus avoiding ELReset errors.

It also brings in a memory leak fix (ATLASRECTS-3925) that was present in MuonTruthAlgs-02-12-16-00 and never made it to release 21.

Finally, it adds the migration of MuonDecayTruthTrajectoryBuilder to const interfaces and of the TrackParticleTruthAlg to DataHandles.
This occurred because the development started in svn and these changes were present in the most up-to-date tags, so they got added by accident.
If not desired, these particular changes can be reverted: however, as they should have no effect on the performance, I left them in.

Tier0 test results can be found at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-3796_git/: no T0 violation was observed.
The only changes that would be involved here are adding dummy values for cases when truth-matching fails.
For the element-links, I don't know if this would be caught by the diff-root.
However, as dummy values are also added for the truthOrigin and truthType, there would be in principle some T0 violation (though for MC only of course).
Since these are dummy values and so will not be used by analyzers, I don't think it's a significant violation.

AODtoTAG also reported many more errors for q221: since there were no changes at all in any of the other substeps, I assume this is a glitch of some sort.


Former-commit-id: e672f80dc1f6a5a94b2ac1ff178859ccb74b9457