adding MS/ID-scatterers for all extrapolated tracks
Currently the xAOD muon holds just for its primary track particle
the values of the MS/ID-scatterers. However, for debugging and
subsequent studies it would be convenient to have this scatterers
for all extrapolated tracks associated to the muon. With this
commit the values are added as auxiliary decoration to those
tracks.


Former-commit-id: 91e3572bffbf256bc54f719af790753ab4bce385