Add a new MonitoringToolBase API define
This is needed to allow better granularity in the change to the interface.
The current API define was added sometime ago when the interface changed.
Now the interface has been changed again, so it is necessary to be able to
differentiate between the *original* 201401 version of the interface, and
the interface as it is defined now.

The old interface API is still kept to preserve backwards compatability.


Former-commit-id: 16e436a5d576f072b791702119efa1011112f217