extrapolation to vertex rather than beamspot for MS part of CB track (ATLASRECTS-3914)
When the MS measurements of a CB track are re-extrapolated to the
production point of the muon after the ID measurements are stripped
off the target was up to now the beamspot. This constraint gave
the ME track quite some freedom in particular if it was a (de facto)
2-station track. With this code the traget is changed to be the
vertex of the ID track which was matched to during combination.
The tighter contraint increases the precision of especially poor
ME track significantly because the vertex acts as an additional,
kind of artificial measurement. The fruther bias of the ME track
by ID information can be accepted since the ME was anyway already
biased by the combination. Moreover, the fully unbiased
extrapolation of the spectrometer information remains untouched
in the MSOnlyExtrapolated track.

A further improvement of the stability of the fit enters with
this request as well: For the extrapolated tracks the ID
measurements got removed while the scatterers where not touched.
This led in some cases to confusions of the fit becaues of
too many freedom. With the new code the ID scatterers are
collapsed to just one scatterer for the extrapolated tracks.
This improves the stability.

The change is frozen tier 0 violating and the corresponding
log files can be found here:

/afs/cern.ch/work/j/jomeyer/public/ATLASRECTS_3914

Like expected only extrapolated tracks are affected.


Former-commit-id: 597e81938405846e52f15a0297476c411c817d6c