decoding and cabling for BMG chambers
During the winter shutdown 2016/1017 ten new small tube MDT
chamber got installed in the feet region (called BMG). Because
of delays in getting the cabling and difficulties to load it
on the database an arbitrary cabling got hardcoded at first.
This code change is adjusting the arbitrary cabling to the
actual cabling (keeping it hardcoded).

Furthermore these BMG chambers are read out by HPTDC chips
rather than AMT chips like all the other MDTs up to now.
Since the MdtAmtHit class suits also to be filled with the
information obtained from the HPTDCs it was kept the same
and only the corresponding readout classes got newly
implemented. The MDT encoder/decoder were adjusted to
use the new readout classes if they hit a BMG chamber.


Former-commit-id: 9020a27788783d030dd07ca3b95d113964e1678c