Modifying the initialization of Chargino processes for ATLASSIM-3231
The SUSY group asked for some tweaks to the chargino process
initialization.  One is to get rid of the addition of transportation,
which simply does not need to be there.  The other is to add a new
ionization and MSC process to each particle, rather than the same one.
Without checking thoroughly, I am guessing there is some caching in the
process that is done, and the results are different because of the
different charges of the particles that its being used for.


Former-commit-id: 1ccb57b8d346fe396c00d69aecec6770ba100754