First version of checkpoint-restart support in AthenaMP and in PyJobTransforms
PyJobTransforms: introduced two new command-line options --checkpoint (bool)
and --restart (string). It is yet to be defined what value we should be passing
to the --restart option

AthenaMP: added code to checkpoint itself just before forking and then to
exit immediately. The code cannot yet be released because it depends on private
installation of DMTCP 2.4.5. Pending the appearance of DMTCP as Athena external.

Former-commit-id: 72d35ff6c66c6671cb88ad32cd771f5f2a380f0d