Add two jet containers for 2017 trigger (ATR-16214)
Two new containers have been added for 2017 trigger usage.  The
SoftKiller container is for trigger studies and thus is only
in AODFULL, while the large-R reclustered container with an
updated calibration is aimed at physics usage and thus is in
AODFULL, AODSLIM, and AODVERYSLIM.  The size of this container
for physics usage is <0.5kb/evt in the standard MC test sample.


Former-commit-id: 43b664ec633161b77562607a6135d9651033cf0c