CSC segment finding if eta and phi layers have different status
The current CSC segment finding assumes that eta and phi layers always have the same status.
That is, if the eta layer is good, the phi layer is also, and vice versa.
However, due to a hardware problem, this will no longer be the case for some CSC chambers this year.
This update modifies the CSC segment-finding code so that layer status is treated separately for eta and phi layers.
The statuses of the eta and phi layers are checked in the Csc2dSegmentMaker and stored separately.
Each status is then evaluated independently in the CscSegmentUtilTool to determine if it makes sense to use the 2-layer segment finding or not.
Thus, 2-layer segment finding could in principle be carried out only for phi, while for eta three or more hits would still be required.
(Instead of automatically doing 2-layer segment finding for both eta and phi, as was done before.)
Furthermore, if all eta(phi) layers have a bad status, a segment can now be built from only the phi(eta) segment, if one was constructed.
But note that this only happens in the case of bad status: if the eta/phi segment is missing but should be there, no final segment is built.
To allow this, the MuonSegmentCombination now knows about the status of all layers in its chamber.
The Csc4dSegmentMaker and CscSegmentUtilTool can then use the status to determine whether to accept an eta- or phi-only segment.


Former-commit-id: 0b9efa6e1952ceb5cd9e416f4690f9f88219268d