Additional jet chains for 2017 menu (ATR-16021)
After the first pass of the rate evaluation, it is clear that some
jet triggers need to have higher thresholds, or at least have the
option for higher thresholds, than was previously expected.  This
commit adds chains to support these higher thresholds.  Most of
the added chains are expected to be kept as backups (not needed
at 2e34), but some are likely to become primaries by 2e34.


Former-commit-id: df91a301658d4276f15c0802974f998d2f1adaf3