Adjust CSC segment quality
If a CSC segment is built only from eta or phi hits because the chamber's phi or eta strips are off or broken, we do not wish this to unduly affect the quality.
In particular, CSC segments with fewer than four eta+phi hits will be given the lowest quality.
However, we could easily have a CSC segment with 3 (or, in rare cases, 2) hits in eta or phi only if only eta or phi is functional.
Therefore, flags are now added to define the eta and phi status for CSC segments.
If the status is bad for eta or phi, a segment with fewer than four hits will be given the next-to-lowest quality.


Former-commit-id: 0905934a5ef3679ecd06c34030ed283894d675d4