Add python access methods to extract info from triggerdb (ATR-15799)
We are in the process of disabling the automated storage of trigger
menu and bunch group information in COOL for online running. Instead
processes will use COOL only for primary keys and then access the
Trigger database directly for the rest of the information. As part of
this migration we are adding extra python-based access methods for
TriggerDb information which are needed to enable various clients
(e.g. datasummary) to migrate from COOL.


Former-commit-id: ce7f1fb75eadf779fbad97cdcf90729b09b60c34