Fix race condition in AtlasPyFwdBwdPorts build
In AthSimulation builds of AtlasPyFwdBwdPorts the installation of
two python packages fails due to unmet dependencies. The fact that
this does not occur in Athena builds suggested a race condition.
The problem packages were determined to be grin and flake8. Adding
approriate DEPENDS statements for these packages solved the issue.
See ATLINFR-1680.


Former-commit-id: 431fcd2e10a1cc58fc31f3a160ac0ab80c267341