Fixed the build of some more minor things in CxxUtils on MacOS.
Tried to first just update the CxxUtils package to its latest version
that's in the master branch, but that would've required updating a whole
lot of other packages as well. So instead I just fixed the existing minor
issues in this code. So that it could even be backported into 21.0.


Former-commit-id: 874a90fcdcf25777246b0f4f7dc23cde0ded543a