Got rid of IGeoModelSvc::geoInit() and cleaned up all of its clients
Long time ago we implemented a mechanism of data-driven geometry initialization,
which was based on TagInfo callbacks. A number of clients was migrated, but this
infrastructure has never been used since that.

In this commit we cleaned up the useless code from core GeoModel and from all
of its clients.


Former-commit-id: 804936490ddd43f969adbc0c455abee0e2183683