Downgraded to ParticleJetTools-00-03-51.
This was needed because the latest updates in AsgTools introduce
const-correctness for ToolHandle-s in standalone mode, and the latest
version of ParticleJetTools only works with the latest version of
MCTruthClassifier in that mode. But we don't use the latest version
of either of these in the 2.6.X nightlies at the moment. (For
MCTruthClassifier we actually use a branch tag.)


Former-commit-id: 7e01df28f721b947de302156d6ef35ff5441bdaa