Fix for ATLASRECTS-3788
This commit fixes ATLASRECTS-3788, addressing an issue in which muon truth auxdata was not always saved to muons.
The code loops over xAOD::Muon objects and truth particles and matches them using the muon primary track particle.
Previously, the loop was over the truth particles and then the muons: I have reversed this.
I feel it makes more sense this way as the muon dummy truth information will then only be set (if necessary) once.
There is no FT0 violation as the only change is the addition of empty vectors to muons that are not truth-matched.


Former-commit-id: 3c51afa2a90447b7da833a70f013d9f009247248