Modification to the ZFinder to increase robustness
Modifications to the ZFinder to allow setting of a maximum layer
to use for creating doublets and triplets to reduce background.
Also changes to reject vertex candidates where the vertex significance
is small.

These changes are needed for the algorithm to be used to seed the
beamspot tracking in high pileup events


Former-commit-id: 042dbe54be13f9c42915777e020696b38be185f2