Adding photon conversion to muons as a particle origin
It has been observed in some new ttbar samples that a significant number
of additional muons are coming from photon conversions to muons in the
truth record.  Presumably this is just a modification to the generator
treatment of vertices that were already allowed.  In order to deal with
this case, muons are now allowed to have a photon conversion origin if
the truth record indicates that is what is going on.


Former-commit-id: d7a5c6f8e57f98e5427028173128d02db0585945