TrigConfReadWrite application to return FATAL if metadata write fails (ATR-16103)
When TrigConfReadWrite is unable to complete populating in-file metadata it
typically aborts in such a way that any reco job in which it is running does
not register the failure. This can then cause confusing crashes in later jobs,
as demonstrated in ATR-16121. This change will cause the tool to issue a 'FATAL'
to the log of the original job, meaning processing will stop at the point of
the original error.


Former-commit-id: 0c0c4e36bfc16f65363c4b016ad88476ba01b2e6