Fixes for running Cavern Background Simulation
Rather than having a ToolHandle to the `TrackFastSimSDTool`, the
`NeutronFastSimTool` is given the name of actual `TrackFastSimSD`
instance to be retrieved. This gets around the issue that the
`TrackFastSimSDTool` is no longer public. It means that the
`NeutronFastSimTool::initialize` method is now redundant also.
Should fix ATLASSIM-2813

Removed `DummyVolumeName` from volume list of `TrackFastSimSDTool`.
The current logic in `SensitiveDetectorBase` with the `noVolumes`
flag is sufficient to ensure the SD is added to the SD manager.


Former-commit-id: 6fb178cb3317e07dddba7f908b5cec5382bc570d