Updated/downgraded to xAODRootAccess-00-01-62-01.
This was not so easy to figure out based on the SVN tags. The tag used
in 2.6.X (xAODRootAccess-00-01-65-01) relies on an older implementation
of the auxiliary store code than what's in 21.2. But the trunk tag
also relies on Gaudi/core sw features that are not part of 21.2.

So this tag should be the right middle ground. It makes use of the latest
AthContainers features, without needing the latest StoreGate updates.


Former-commit-id: 7edcff3e7a71d0e14e29d80c0d649af6fb7f6aa8