partial fix for very rare crashes (ATLASRECTS-4005)
In the reprocessing of 2015/2016 data there had been 5 crashes caused
by muon combined reconstruction. Four of them are fixed by this merge
request. It is not violating frozen tier 0 as demonstrated with
/afs/cern.ch/work/j/jomeyer/public/RunTier0Tests.log_ATLASRECTS-4005
However, since it fixes very rare crashes it sometimes changes the
output. The fix adds a protection preventing the iteration beyond
the end of a vector.


Former-commit-id: a53fd32d5e50702aae62ca24b90bea34c66249a0