AthContainers+CxxUtils: Fix ubsan warnings.
A (deliberate) reinterpret_cast in copyAuxStoreThinned triggers a warning
from the undefined behavior sanitizer in gcc6.  For now, disable warnings
for this particular function.  This is done with a gcc-specific function
attribute, so add a macro in CxxUtils to make this visible only for gcc6
or later.

Also fix another ubsan warning in one of the tests.


Former-commit-id: b4de894f9f8b728b067e1a5ab528b7d48a83d2d2