Fix for calo tag chains.
This fixes some double counting in multi-muon chains between calotag and combined muons.
Before this dimuon chains with calotag muons were not guaranteed to have two different muons as the calo tag and combined
muon finding is independent. Some overlap removal was added to the calotag hypo to avoid this.
ATR-16248.


Former-commit-id: 63caafeeac49b3ea75e17b56788b8ff1af8f287c