DerivationFrameworkMCTruth: Fix compilation warning.
Code has a couple instances of `if (!x>0)' which was intended to be true
if x is 0.  It actually does that, but kind of by chance.  The expression
is actually parsed as (!x)>0, so the LHS of this comparison is 1 if x
is 0 and 0 if it is not.  Rewrite this as simply `x==0', which is clear
and unambiguous,


Former-commit-id: 576274b87f5214811e5af1ad01d883c65b9618f1