Some technical fixes for running han-config-gen commands during the build.
Updated the configuration to set up the dependencies of the file
generation custom command correctly. Using file(GLOB ...).

Hid the generated files under ${CMAKE_FILES_DIRECTORY} in the
binary directory, to be in line with how the rest of the CMake code
tries to operate.

All of this also fixed the build of the DataQualityConfigurations
with Ninja.


Former-commit-id: 5bff2e9e6b96a2c2b1fa48c1b53ee1ff41fa609c