Fix for cosmic timeouts in muon combined reco
The timeouts occur when the TrackParticleCreatorTool attempts to extrapolate the muon's primary track to the beamline.
Previously, the TrackParticleCreatorTool extrapolated tracks to (0,0,0) instead.
This makes more sense for cosmics anyway, since there is no beamline.
Changing the configuration of the muon combined reconstruction instance of the TrackParticleCreatorTool to use the origin instead fixes the problem.
Therefore there is no necessity to turn off the muon combined reconstruction.


Former-commit-id: 7b50f6a67f5943064e2863cec09e07ca8c016e6b