Fix parsing of postproc step list
Sometimes the list of postprocessing steps was comma-separated and sometimes it
was space-separated. The argument as passed in would then be stored in the
beamspot task database without validation, implicitly assuming that spaces were
delimiting the steps. Commas would silently be accepted but then when the
string is read out of the database later it is split only on spaces, causing
the job to crash with an unknown step name.

With this change the option parser is extended with a new field that
understands both comma and space delimiting so that the option is stored as
a python list. This list is then uniformly serialised with spaces when passed
to the database logic.

This is part of svn changeset 786208:

  https://svnweb.cern.ch/trac/atlasoff/changeset/786208/InnerDetector/InDetExample/InDetBeamSpotExample/trunk/bin


Former-commit-id: 31d54d3780b7943c4e34f10e5d8504f4b7416b27