Remove svn Id lines
The __version__ variable in each file is used in log messages to explicitly
note the source version running. Since these are no longer updated in the git
world I have given them dummy values for now so as not to be misleading. In the
future this feature may be removed.


Former-commit-id: 9ebe7e4d5326aeee8666c8edfc470e0cb6e39b79