Changes for migrating to the dynamic folder scheme in the online and avoid the need of the conversion from the new scheme to the old.
The scheme used is controlled by
InDetGMProperties.useDynamicAlignFolders setting. This replaces the
old setting of the same name in InDetFlags/InDetTrigFlags and the need
of modifying the offline IndetFlags setting in the online job. It
allows also to simplify InDetAlignFolders and avoid dependency of this
package on the reconstruction. This is related to the discussion in ATLASSIM-2746 and ATR-13872

Further changes are to propagate the setting from *DetectorTool
properties to the actual geomodel DetectorFactories where different
sets of folders are accessed. It allows to deprecate the method
getAlignFolderType which did guessing of the alignment scheme based on
the presence of the folders and which is not robust in the case 2 sets
of folders are present in the db. It also reduces the amount of
WARNINGs produced (ATLASRECTS-3800) and makes them more meaningful.

I'm happy to iterate on the implementation especially because I did
not test but the online usecase (with a combination of the online
database and an sqlite file with the new folders) and for example more
clients of the old InDetFlags setting may need adjustment. The request
is for the moment for the online 21.1 release and I'm marking it
urgent to be able to validate it in a release before the first stable beams.


Former-commit-id: ec81955661102b0fc80ab206ebd5d262f5324854