Make TaskManager a context manager
We ran into a compound issue when running beamspot jobs on the batch system.
The underlying issue was that there is a ROOT bug triggered when the end of the
python script is reached and ROOT enters a cleanup routine. It would crash and
then cause the scripts to hang.

The problem was that we were opening an Oracle connection to the beamspot task
database but deferring the closure of the connection until __del__. Because of
the ROOT bug python never got around to cleaning the TaskManager and so the
connection would remain open. Other jobs in the same batch would then hang
waiting for the database lock to be released causing a deadlock. The jobs would
fill up all available execution slots on the batch queue. To make matters even
worse the default batch queue was atlasb1 which is used for time-critical jobs.

This change instead makes TaskManager a python context manager so that it can
be called using a pattern like:

    with TaskManager(...) as taskman:
      # do something ...

and the connection will be closed at the end of the block. A variable is added
to keep track of cases when this pattern is not being used so that a warning
can be printed. This should be sufficient to minimise deadlock issues like we
observed before.


Former-commit-id: 2c077a1af9f0b523fdbc31c46b5c054a4e6c746d