Make use of the new TaskManager interface
The previous commit introduced a new interface to TaskManager where using it as
a context manager cleans up the database connection at the end of the block.
This commit switches all current uses of the TaskManager to this new interface,
in some cases re-ordering parts of the code locally so as to make the block
with the database connection as minimal as possible.

The majority of lines touched here are only because they needed to be indented
to sit inside a new context block.


Former-commit-id: c117b451f8b7db3db99b46bd71d63e6cb9236d6b