Get rid of IBL flags in NoiseMapBuilder
With this commit backwards compatibility with Run-1 is broken, we now
assume IBL is there. This simplifies the structure of the code of the
NoiseMapBuilder, since we can just stay with a single mapping file
(PixelMapping_Run2.dat), and simplify the logic while iterating
over the pixel barrel layers: layer 0 is IBL, layer 1 is Blayer, etc.


Former-commit-id: c7c06e47a131533467d91fd0565810d83d52c80f