More functionality added to AthenaMP afterRestart() function
Now it is possible to update output file name, max events and skip
events. For updating output file name, AthenaMP fires a new incident
which is then handled by AthenaOutputStream. This might not be the
best solution for implementing this functionality. To be discussed
further.

It was also necessary to make changes in SharedEvtQueueProvider in
order to be able to update the skip events value. The changes are
mostly cosmetic. SharedEvtQueueProvider now retrieves skip event
number from its exec_func() instead of the initialize()


Former-commit-id: 3b36afa026b3086f218e7ecfecba6abe0dc1fea5