Correct logic concerning the track parameters to be converted.
Originally the first and last track parameter set was added to xAOD
TrackParticles if the property KeepParameters was true, and all track
parameter sets were written only, when KeepParameters was not set true
but the property KeepAllPerigee was true. After the inclusion of the
bad cluster identification, all parameters are stored if the
BadClusterID property is false and KeepAllPerigee is true, independent
of KeepParameters.

Now the original behaviour is restored that either the first and the
last or all track parameters are stored depending on the properties
KeepParameters and KeepAllPerigee.


Former-commit-id: 6bfe1887411f5892e8124dd65fdb54506773cb53