Simplify registerHistograms in NoiseMapBuilder
Simplify registerHistograms() in NoiseMapBuilder.cxx, by encoding in the
new private function histoSuffix the histogram names depending if
barrel/endcap and layer. Still the histogram binning is cumbersome
to deal with but I think it's clear enough. Left a lot of code commented
out in case need to refer back to original.


Former-commit-id: edbc0d65eb2fd1c57a2b15a2689affa052195b67