Fix for extremely rare bug in muon extrapolation
In one case in the 2015-2016 dataset, the standalone muon fit was so bad that it appeared to have no hits in the MS.
This causes a problem for the loop in the standaloneFit method of CombinedMuonTrackBuilder.cxx that is supposed to find the first MS hit on the track.
Instead of ending up with an iterator that points to a hit, it has an iterator which points to the end of the track's TSOS vector.
There is an if statement immediately following the loop which is supposed to catch this, but it first checks to see if the hit pointed to has valid track parameters.
In the case that the iterator points to the end of the vector, this causes a crash.
Changing the order of the conditional statements in the if statement removes the crash.
The underlying cause of the poor fit is still under investigation.


Former-commit-id: d703e3a20d91bcaff6cb6139dc2499a952792de3