AnalysisTriggerTools: Update use of obsolete method of TrigMuonEFInfo.
The CombinedTrack method of TrigMuonEFInfo is non-const, and can in fact
change the object.  It was being called on an object from a const DataVector.
The only reason this worked was due to a long-standing DataVector bug
that allowed retrieving a non-const pointer from a const_iterator.

Updated to what appears to be the current way of doing this.


Former-commit-id: 5c6827df1ba00ab003cb01f4f32a0df0a08bfe6f