TgcRawDataMonitoring: Remove some superfluous clears.
Here:

```
   for(int sta=0;sta<4;sta++){// station
     for(int eta=0;eta<6;eta++){// eta
       if(sta==0 && eta==5){
-        tgcsummaryoflog10wireoccupancyperchambertype[ntype] = 0;
-        tgcsummaryoflog10stripoccupancyperchambertype[ntype] = 0;
         continue;
         
       }
       if(sta==3 && eta >1){
-        tgcsummaryoflog10wireoccupancyperchambertype[ntype] = 0;
-        tgcsummaryoflog10stripoccupancyperchambertype[ntype] = 0;
         continue;
       }
       ...
       ntype++;
     }
   }
```

For the first of these, the element at ntype will anyway get filled 
in on the next iteration, so the clear is redundant.  For the second,
we actually hit this case at the end of the iteration, so at that point,
ntype is already pointing past the end of the array, resulting
in an out-of-bounds array write.  (Caught by the undefined behavior
sanitizer.)



Former-commit-id: ed9ffa1939ef46d4afa91b06027e3b16054b1690