ATR-16357: MC_PhaseII menu added
Added all relevant files to add a new trigger menu filed called MC_PhaseII.py which is based on the current v7 L1 menu.
In order to get this to work, e.g. TriggerConfigL1Topo.py and l1menu/ItemDef.py had to be touched to reflect the name
changes as several flags (if '_v7' in triggermenuname then blabla) are used in TriggerMenu.


Former-commit-id: 7c6bb5cba1882b7d1efa2a7360b2a6f0ac1a295f