TrigInDetAnalysis: Fix use-after-delete.
The non-const vector accessors should return the result by reference, not by value. Besides
avoiding a copy, there is code in TrigInDetAnalysisExamples that was doing something like
```
  TIDA::Track* t = &roi.tracks().at(i);
```
But if tracks is returned by value, then we're taking the address of a temporary which becomes
invalid immediately after this statement finishes and the temporary vector is destroyed.
This was triggering a warning from the undefined behavior sanitizer.


Former-commit-id: 975326aa39cb7eb0f75adf37d017a6ade84e31fe