skipping non-existing BMG tubes in DQ (fix for ATLASRECTS-4058)
The newly installed BMG chambers have cutouts, i.e. in some places
there are tubes missing even though their identifiers exist. The
DQ code iterates on the identifiers and asks for each tube the
geometry for the position. Because for non-existing tubes this
position cannot be evaluated an ERROR occurs. With this commit
a member variable is introduced which at the beginning of the
DQ code collects all identifiers of non-existing tubes and later
on in the loop those tubes are skipped.


Former-commit-id: 9b578920aaa77386692c69439c2310728127aa70