Fix GetRelease in JobRunner
I don't see how the old behaviour ever worked, since the test string could
never be equal to the empty string so that if statement would always be true.
However since the code was always run in environments with a patch version the
bug was never noticed.

With the newer cmake+git releases this buggy code path is triggered, so
I instead remove it. Also `os.path.expandvars()` was really being abused here
since if the variable is not set it returns the variable name literally in the
result whereas I think the author expected it to return an empty string.
Instead `os.getenv()` is more appropriate.


Former-commit-id: 3d4467885b3e4ad499565a305b3f830dc57d42d1