StoreGate: Avoid binding to proxies handles created from keys.
When a handle is created, it binds to the associated proxy, and then unbinds
when the handle is deleted.  This is needed to handle objects that persist
across events.  However, for handles which are created as local variables,
this is unneeded overhead.

Avoid binding/unbinding handles initialized from key objects, since the
intention is that those are used locally.


Former-commit-id: 3b6f5fef3f9a3a0765f6c904d6170ff79b53d9b5