L1Topo: add comparison of DAQ TOBs vs. ROI TOBs
Details:
This comparison helps detect events in which there is a mismatch
between the readout of the L1Topo CTP bits in the main "DAQ" (ROS)
readout which also has the inputs TOBs and the smaller fragment we
send to the RoIB which just contains the CTP bits.

The comparison of all the words is non-trivial, because of different
ordering and zero suppression (that is, some TOBs could be missing
from the DAQ output). So what is being done here is that I compare
only the 128 decision bits (the overflow bitset is also computed, but
not being compared for now).

This is done by taking the L1Topo TOBs for BC0 from the DAQ ROB, which
are available in `daqTobsBC0`, and map them on a bitset[128] in the
same way being done for `m_triggerBits`, which already contains the
trigger bits from the ROI TOBs.

This code has been tested with RAW files from 2016 (run 311071, in
which the readout was working properly) and RAW files from 2017 (run
324320, for which @landon reported mismatches).


Former-commit-id: 6d4590e9eb41590955822774928e5da159087a67