Modifications to fis the bjet vertex tracking threshold
Originall, all analyses in group needed to use the same pt threshold.
These changes allow the setting of individual thresholds for each analysis,
which is essential for the bjet vertex tracking which now uses 5 GeV as
a threshold for the tracks.

Therefore with these changes we can monitor the efficiency of the bjet
vertex tracking only for tracks about 5 GeV, while retaining a lower
threshold for the other chains.


Former-commit-id: dbd1829c52e6975097664ccfb82cfa159ba3e5b1