Fix for ATLASRECTS-4060
This fixes an occasional crash when reconstructing toroid-off data.
The problem turned out to be that the standaloneRefit method is not protected against the 0 magnetic field case.
Since the refit is not really necessary for processing toroid-off data, the simplest solution is to turn it off.
Since this fix only affects toroid-off data, no FT0 violation is expected or observed.
Results of running the Tier0 tests can be found at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-4060/.
(Note that although the log indicates there were changes in MuonCombinedBaseTools, there were none, I simply checked out the package by mistake and forgot to remove it.)


Former-commit-id: c22871d92733ed2d9b0aae810312e99112d5a0d2