Fix for ATLASRECTS-3990 to avoid infinite loop
There is a while loop (starting in line 1833) which will never break under certain conditions because it keeps iterating on the same object since some if-statements miss to increase the iterator before continuing the loop. In production this issue may not have been seen, however, in private processings of merged single RAW events it appears. Most likely this is because such processing could have a different call-back sequence of the conditions service than the bulk processing.

Former-commit-id: 9c232e948d01fb81ecbf5a9201c4844f6022d591