SCT_RawDataByteStreamCnv: Fixes for recording BS error object.
The SCT_RodDecoder tool was creating and recording a BS error object
in a BeginEvent incident handler, then writing to it later.
This is not MT compatible.
Rework so that the error object is created in the algorithm that uses
this tool, at the same time as the RDO container, and is then passed
through the interfaces to the tool.

Some other cleanups as well.


Former-commit-id: a0a48a27f38c97cc600807973bd44830bf794049