Replace std::map::insert() with direct assignment
As discussed on ATEAM-402, attempting to modify conditions data that is stored in a map using map::insert() will fail.
The insert() method only adds information to the map if the key does not already exist: if the key exists, it does nothing.
This was originally investigated for the SCT conditions, but the muon conditions also has this problem in several places.
Fixing this should have a minimal effect on the final result, as the status information is used only in the hole search.
So only the number of holes on track should be affected, with possible tiny changes from the addition of a hit to a track.
(This could happen if the conditions was not correctly updated to indicate that the multilayer it is in was actually working.)
FT0 violation should be quite small, and indeed none is visible in the FT0 tests, available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATEAM-402/.


Former-commit-id: d143b51341e243888e43902913b0c6a429587677