Fix for combined muon reconstruction in cosmics
For some reason cherry-picking MR 1743 directly didn't work so I am doing it manually.
Please see that MR for more details: this should not be particularly important for the trigger, but it keeps the online and offline codes consistent.
As it has been in 21.0 for a couple of weeks, this change has already passed RTT tests equivalent to trigtest.pl.


Former-commit-id: bc043ee6f7d31fb0316ee3bff1907c7bfc75075f