AthenaMonitoring: Fix crash when exiting test.
GenericMonFilling_test was not finalizing THistSvc.
This means that the output file will get closed while global dtors are running.
Depending on the order in which they run, this can lead to crashes.
(In particular, it was observed that cling would sometimes get cleaned
up before the files get closed, leading to crashes when the I/O code
tries to use TClass interfaces.)

Explicitly finalize services before exiting.


Former-commit-id: 58abb782dbde9992a514b1b8aa6d9ff3971dd0fc