Pythia_i: Make sure that pydata.o gets linked into Pythia_iLib.
For pythia to work, the separate object pydata.o needs to be linked
along with the pythia shared library itself.  In the cmt build, this
was added via the requirements of External/Pythia; however, that's
no longer done for the cmake builds.  So link the object from here.

This is one of the changes needed to fix ATN failures in MCParticleTests.


Former-commit-id: 277df813d28c6cd6879faf6f1795f888535da04f