add DAQ_ROB_Hdw_vs_CTP_Hdw_Events comparison
Details:
this is an additional monitoring that we need while the ROIB readout
is scrambled. It compares the trigger bits from L1Topo with the ones
at the CTP input. The L1Topo bits are OR'ed with the overflow bits, in
the same way as already done for the ROIB bits in the histogram
`CTP_Hdw_vs_L1Topo_Hdw_Events`.

Former-commit-id: 55a6854ec52087bf408342b8a8b0a89c5d292e86