Allow possibility to select on number of si holes in the ID Trigger Tier 0 monitoring
Have job options to set the number of SCT and pixel holes, but these
are not used to filter the tracks.

These changes use the values set by the job options to filter the tracks.
The values are the current defaults, so no differences should be observed
unless these job options are explicitly set differently


Former-commit-id: 22b7c0ba0185e8f1943f0fec9286f7f0cccf4417