Fix issue in RunTier0Tests when looking for most recent nightly
The past implementation was such that the check for the most recent
nightly would fail if RunTier0Tests.py is run on top of a nightly directly
after an update to the version.txt file (the "*" wildcarding would
not resolve to multiple paths, and so the output of the previous `ls`
command would be different than the code was expecting.  This new
implementation has been confirmed to fix this issue.


Former-commit-id: 9f058a27103705b9de3e71533bd4e012bb7be7d0