Remove offline beamspot cron scripts
These scripts are run within ~atlidbs/cron to perform various beamspot tasks
including within the calibration loop. They don't need to be in a release since
the only place they are used is a local checkout in ~atlidbs. For this reason
I've moved them to a separate repository at:

    https://gitlab.cern.ch/atlas-beamspot/beamspot-cron

The versions there are the ones currently in use.


Former-commit-id: f772d6306c6e8eee25a77b0372efc1cc03074329