Added the CaloClusterCellLinkContainer for standard egamma clusters.
This was previously done with CaloClusterStoreHelper::finalizeClusters, but
that didn't work with write handles. Therefore, the pertinent code was added
in the EMClusterTool.

Note:  CaloClusterStoreHelper::finalizeClusters was removed everywhere,
but this code was only added back in standard egamma. Forward electrons
still need to have this added. (Forwad electrons haven't been migrated to
handles yet.)

Former-commit-id: 3c46600d105ab9ca04605ea98ff0ece025154e5b