Fix for ATLASRECTS-3199
This fixes a bug where position and direction information was not filled for truth muon segments built from CSC hits.
CSC sim hits are treated differently than MDT ones, and it seems the code to handle them correctly was simply never written.
The fixes in MuonTruthDecorationAlg::createSegments simply calculate this information for CSC segments the same way it is done for the other segments.
The result will be FT0-violating, though only for MC, of course.
Note that this violation is not visible in the Tier0 tests as q221 happens to not reconstruct any truth CSC segments.
Tier0 test results are available at /afs/cern.ch/work/w/wleight/public/MuonSW/ATLASRECTS-3199/.


Former-commit-id: 41c73845fa7d10248228f1b5fb0e4fbc8356b6b8